change CheckAndObfuscate.h
